ScheduleBuilder/requirements.txt
sqlparse==0.4.4
asgiref==3.7.2
Django==4.2.5
django-environ==0.11.2
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
python-decouple==3.8
PyPDF2==3.0.1
pandas==2.1.1
nltk==3.8.1
python-dateutil==2.8.2
pdfplumber==0.10.2
django-allauth==0.57.0
dj-rest-auth==5.0.1
google==3.0.0
# Not sure if these are required, I'll get rid of them if they're not
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
httplib2==0.22.0 
idna==3.4 
oauthlib==3.2.2 
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0 
pyparsing==3.1.1
requests==2.31.0 
requests-oauthlib==1.3.1 
rsa==4.9
uritemplate==4.1.1 
urllib3==1.26.16
